cha: re-structure CLI options to use subcommands
