+++ src/nexuscli/cli/errors.py
@@ -1,0 +1,7 @@
+from enum import Enum
+
+
+class CliReturnCode(Enum):
+    SUCCESS = 0
+    INVALID_SUBCOMMAND = 10
+    POLICY_NOT_FOUND = 20
