+++ src/nexuscli/cli/subcommand_cleanup_policy.py
@@ -1,0 +1,69 @@
+"""Nexus 3 CLI
+
+Usage:
+  nexus3 cleanup_policy --help
+  nexus3 cleanup_policy create <policy_name> [--format=<format>]
+         [--downloaded=<days>] [--updated=<days>]
+  nexus3 cleanup_policy list
+
+Options:
+  -h --help             This screen
+  --format=<format>     Accepted: all or a repository format [default: all]
+  --downloaded=<days>   Cleanup criteria; last downloaded in this many days.
+  --updated=<days>      Cleanup criteria; last updated in this many days.
+
+
+Commands:
+  cleanup_policy create  Create or update the cleanup policy <policy_name>
+  cleanup_policy list    List all existing cleanup policies.
+"""
+from docopt import docopt
+from texttable import Texttable
+
+from .. import cleanup_policy
+from .errors import CliReturnCode
+from .util import find_cmd_method, get_client, TTY_MAX_WIDTH
+
+
+def cmd_list(nexus_client, args):
+    policies = nexus_client.cleanup_policies.list()
+    if len(policies) == 0:
+        return CliReturnCode.POLICY_NOT_FOUND.value
+
+    table = Texttable(max_width=TTY_MAX_WIDTH)
+    table.add_row(['Name', 'Format', 'lastDownloaded', 'lastBlobUpdated'])
+    table.set_deco(Texttable.HEADER)
+    for policy in policies:
+        p = policy.configuration
+        table.add_row([
+            p['name'], p['format'],
+            p['criteria'].get('lastDownloaded', 'null'),
+            p['criteria'].get('lastBlobUpdated', 'null')])
+
+    print(table.draw())
+    return CliReturnCode.SUCCESS.value
+
+
+def cmd_create(nexus_client, args):
+    criteria = {}
+    if args.get('--downloaded'):
+        criteria['lastDownloaded'] = args.get('--downloaded')
+    if args.get('--updated'):
+        criteria['lastBlobUpdated'] = args.get('--updated')
+
+    policy = cleanup_policy.CleanupPolicy(
+        None,
+        name=args.get('<policy_name>'),
+        format=args.get('--format'),
+        mode='delete',
+        criteria=criteria,
+    )
+
+    nexus_client.cleanup_policies.create_or_update(policy)
+    return CliReturnCode.SUCCESS.value
+
+
+def main(argv=None):
+    arguments = docopt(__doc__, argv=argv)
+    command_method = find_cmd_method(arguments, globals())
+    return command_method(get_client(), arguments)
