+++ tests/cli/test_delete.py
