+++ tests/cli/test_download.py
@@ -129,7 +129,7 @@ def test_download_tree(
     dst_dir = faker.uri_path() + '/'
     path = dst_dir[:-1] + src_dir
 
-    argv = ('repo create hosted raw {}'.format(repo)).split()
+    argv = ('repository create hosted raw {}'.format(repo)).split()
     pytest.helpers.create_and_inspect(nexus_client, argv, repo)
     nexus_client.repositories.refresh()
