+++ tests/cli/test_list.py
